Extract track view and add preview data
Using a `Track` object to manage the viewed data allows us to easily
extract a subview that takes a `Track` as its sole argument.

To aid in the Xcode editor, we add a `Track.example` object that reads
from a sample JSON file. Both the property and the source JSON will not
be included in the compiled executable, as the `Preview Content`
folder is in the project's "Development Assets" list.